+++ domvm-v2.1.3-keyed/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
 <head>
   <meta charset="UTF-8">
-  <title>domvm v2.0.3-non-keyed</title>
+  <title>domvm v2.1.3-keyed</title>
   <link href="/css/currentStyle.css" rel="stylesheet"/>
 </head>
 <body>
