+++ domvm-v2.1.3-keyed/src/store.es6.js
