+++ domvm-v2.1.3-non-keyed/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
 <head>
   <meta charset="UTF-8">
-  <title>domvm v2.0.3-keyed</title>
+  <title>domvm v2.1.3-non-keyed</title>
   <link href="/css/currentStyle.css" rel="stylesheet"/>
 </head>
 <body>
