+++ domvm-v2.1.3-non-keyed/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "js-framework-benchmark-domvm-keyed",
-  "version": "2.0.1-keyed",
-  "description": "Benchmark for domvm framework (keyed)",
+  "name": "js-framework-benchmark-domvm-non-keyed",
+  "version": "2.1.3-non-keyed",
+  "description": "Benchmark for domvm framework (non-keyed)",
   "scripts": {
     "build-dev": "node build.js",
     "build-prod": "node build.js"
@@ -13,6 +13,6 @@
     "rollup-plugin-uglify": "*"
   },
   "dependencies": {
-    "domvm": "git://github.com/leeoniya/domvm.git#8ab7a791f43d4b0632af15a8e39ad8c199c2fd9a"
+    "domvm": "git://github.com/leeoniya/domvm.git#0c926f2b9055c8a15c3189c01cbe8a8523473e7f"
   }
 }
