+++ domvm-v2.1.3-non-keyed/src/store.es6.js
