+++ webdriver-ts/src/common.ts
@@ -43,8 +43,8 @@ export let frameworks = [
     f("choo-v4.1.0", true),
     f("cyclejs-dom-v14.1.0", true),
     f("dio-v3.0.5", true),
-    f("domvm-v2.0.3-non-keyed", true),
-    f("domvm-v2.0.3-keyed", false),
+    f("domvm-v2.1.3-non-keyed", true),
+    f("domvm-v2.1.3-keyed", false),
     f("ember-v2.6.1", false, {uri: "ember-v2.6.1/dist"}),
     f("ember-v2.10.0-beta.3", false, {uri: "ember-v2.10.0-beta.3/dist"}),
     f("elm-v0.18.0", false),
